modified:   index.html 	modified:   script.js 	modified:   style.css
A lot of modification has been made to make the code function has expected.
However, when I click on transparent button, the program does not really function as expected.
The program returns shadeColor = #000000NaN as hex color and I am yet to figure out why.
The most probable reason for this is that, the RGBA code might be undefined; so when I try to convert it to hex, it causes problems