modified:   index.html 	modified:   script.js 	modified:   style.css
The transparent button now functions as expected because I made the program to only use rgba colors and set if statement to handle undefined or not found alpha values

I also add a random color button so that when I click on it, I get random rgb colors.

The random color button functions wells, but a problem arise when I transition to random color button after using the transparent button. At this level, the program does not function as expect.