modified:   script.js
Previously, when I toggle between the transparent and random color button, the program did not function as expected.
So, in the addBackgroundColor() function, I include removeEventListener on all div to make sure all events currently going on are removed before adding a new one.
And Voila.... my program works as expected.